+++ my-app/package.json
@@ -20,9 +20,12 @@
     "axios": "^1.4.0",
     "framer-motion": "^10.12.17",
     "officeparser": "^3.3.0",
+    "pptx-parser": "^1.1.7-beta.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "react-speech-kit": "^3.0.1",
+    "react-speech-recognition": "^3.10.0",
     "recoil": "^0.7.7",
     "styled-components": "^6.0.1",
     "typescript": "^4.9.5",
